+++ termal_core/src/raw/sys/mod.rs
@@ -1,6 +1,6 @@
 use std::time::Duration;
 
-use crate::error::{Error, Result};
+use crate::{Error, Result};
 
 #[cfg(unix)]
 mod unix;