+++ termal_core/src/raw/wait_for_in.rs
@@ -3,7 +3,7 @@ use std::{
     time::Duration,
 };
 
-use crate::error::Result;
+use crate::Result;
 
 use super::wait_for_stdin;